Re-arranged some of the code *NOT TESTED PROPERLY
Need to add the log integration
