+++ go.mod
@@ -7,5 +7,4 @@ require github.com/machinebox/graphql v0.2.2
 require (
 	github.com/matryer/is v1.4.0 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
-	github.com/scizorman/go-ndjson v0.0.0-20200902005011-1d92486df71e // indirect
 )
